move init methods to utils
