+++ ast_engine/config/registry/enrichment.py
@@ -1,22 +1,31 @@
 import logging
-from models import BaseDataset, RegistryDataset
+from .models import BaseDataset, RegistryDataset
 from typing import Optional
 from pathlib import Path
+import uuid
 
 logger = logging.getLogger(__name__)
 
 class Enrich():
+    '''Enrichment class to enrich BaseDataset with additional metadata
+    TODO: might be nice if enrich had parameters (version: str, dataset_list: List)
+    amd returned a Registry
+    '''
     FILE_TYPES=['.gdb','.shp']
     def __init__(self, base: BaseDataset):
         # enrichment state
+        self.id: Optional[str]
         self.columns: Optional[list[str]] = None
         self.geom_column: Optional[str] = None
         self.geometry_type: Optional[str] = None
         self.crs: Optional[str] = None
         self.data_adapter: Optional[str] = None
+        self.row_count: Optional[int] = None
+        self.base = base
     def resolve_adapter(self):
+        # resolve the data adapter based on the datasource
         if self.base.datasource.upper().startswith("WHSE"):
-            self.data_adapter = "ORACLE"
+            return "ORACLE"
         else:
             ds =Path(self.base.datasource)
             if ds.suffix.lower() in self.FILE_TYPES:
@@ -32,14 +41,19 @@ class Enrich():
         self.crs = 'EPSG:3005'
         self.geom_column = 'GEOMETRY'
         self.geometry_type = 'POLYGON'
+        self.row_count = 200
     def enrich_from_oracle(self):
         # use the oracle data adapter to get this info
         self.columns = ['FOO','BAR']
         self.crs = 'EPSG:3005'
         self.geom_column = 'GEOMETRY'
         self.geometry_type = 'POLYGON'
+        self.row_count = 400
     def enrich(self):
         '''Resolves data adapter and enriches object with metadata'''
+        # TODO: Do we need to somehow enforce unique? or move this up to the 
+        # hydration level or Registry object level to ensure unique
+        self.id = str(uuid.uuid4())
         self.data_adapter = self.resolve_adapter()
         if self.data_adapter == 'FILE':
             self.enrich_from_file()
@@ -54,11 +68,13 @@ class Enrich():
         
         return RegistryDataset(
             **self.base.model_dump(),
+            id=self.id,
             columns=self.columns,
             geom_column=self.geom_column,
             geometry_type=self.geometry_type,
             crs=self.crs,
             data_adapter=self.data_adapter,
+            row_count=self.row_count,
         )
